// src/app/intranet/constants/intranet.enum.ts
export enum Intranet {
  BuscadorApi = 'buscadorp/',
  VtaqApi = 'api-sga-bj/',
  HistorialApi = 'historial/',
  ArchivosApi = 'archivos/',
  Portada = 'portada/',
  SugerenciasApi = 'sug-docs/',
  DatosAbiertos = 'datos-abiertos/',
  AutoCompletadoApi = 'autocompletado/',
  SiteExpedientes = 'Tematica/Detalle.aspx?AsuntoID=',
  DescargaArchivoSentencia = 'Tematica/RespondeFileEngrose.aspx?Id=',
  Socket = 'websocket/socket',
  BuscadorBjdh = 'buscador/',
  BuscadorOnu = 'buscadornu/',
  Portafolio = 'portafolio/',
  InformationSource = 'information-source.json',
  SugSearch = 'busquedas-@.json',
  DataSources = 'data-sources.json',
  Guide = 'data-guide-int.json',
  BuscadorApiPub = 'api/buscador/'
}